fix bug

+++ AutoCkt/Shared/shared/typing.py
@@ -1,1 +1,2 @@
-Number = int | float
+# FIXME everything converts to float, why?
+Number = float | int
